refactor(tools/langfuse-export): group src into clients/ (I/O) + pipeline/ (pure) [NES-1690]
